pyproject.toml
[project]
name = "dota-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.68",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
